+++ GoExtractor/out.txt
@@ -1,0 +1,66 @@
+add,(Ident)^(FuncDecl)_(FuncType)_(FieldList)_(Field)_(Ident),x
+add,(Ident)^(FuncDecl)_(FuncType)_(FieldList)_(Field)_(Ident),y
+add,(Ident)^(FuncDecl)_(FuncType)_(FieldList)_(Field)_(Ident),int
+add,(Ident)^(FuncDecl)_(FuncType)_(FieldList)_(Field)_(Ident),int
+add,(Ident)^(FuncDecl)_(BlockStmt)_(DeclStmt)_(GenDecl)_(ValueSpec)_(Ident),z
+add,(Ident)^(FuncDecl)_(BlockStmt)_(DeclStmt)_(GenDecl)_(ValueSpec)_(Ident),int
+add,(Ident)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(Ident),z
+add,(Ident)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),x
+add,(Ident)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),y
+add,(Ident)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(BasicLit),2
+add,(Ident)^(FuncDecl)_(BlockStmt)_(ReturnStmt)_(Ident),z
+x,(Ident)^(Field)_(Ident),y
+x,(Ident)^(Field)_(Ident),int
+x,(Ident)^(Field)^(FieldList)^(FuncType)_(FieldList)_(Field)_(Ident),int
+x,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(DeclStmt)_(GenDecl)_(ValueSpec)_(Ident),z
+x,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(DeclStmt)_(GenDecl)_(ValueSpec)_(Ident),int
+x,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(Ident),z
+x,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),x
+x,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),y
+x,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(BasicLit),2
+x,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(ReturnStmt)_(Ident),z
+y,(Ident)^(Field)_(Ident),int
+y,(Ident)^(Field)^(FieldList)^(FuncType)_(FieldList)_(Field)_(Ident),int
+y,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(DeclStmt)_(GenDecl)_(ValueSpec)_(Ident),z
+y,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(DeclStmt)_(GenDecl)_(ValueSpec)_(Ident),int
+y,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(Ident),z
+y,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),x
+y,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),y
+y,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(BasicLit),2
+y,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(ReturnStmt)_(Ident),z
+int,(Ident)^(Field)^(FieldList)^(FuncType)_(FieldList)_(Field)_(Ident),int
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(DeclStmt)_(GenDecl)_(ValueSpec)_(Ident),z
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(DeclStmt)_(GenDecl)_(ValueSpec)_(Ident),int
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(Ident),z
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),x
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),y
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(BasicLit),2
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(ReturnStmt)_(Ident),z
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(DeclStmt)_(GenDecl)_(ValueSpec)_(Ident),z
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(DeclStmt)_(GenDecl)_(ValueSpec)_(Ident),int
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(Ident),z
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),x
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),y
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(BasicLit),2
+int,(Ident)^(Field)^(FieldList)^(FuncType)^(FuncDecl)_(BlockStmt)_(ReturnStmt)_(Ident),z
+z,(Ident)^(ValueSpec)_(Ident),int
+z,(Ident)^(ValueSpec)^(GenDecl)^(DeclStmt)^(BlockStmt)_(AssignStmt:=)_(Ident),z
+z,(Ident)^(ValueSpec)^(GenDecl)^(DeclStmt)^(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),x
+z,(Ident)^(ValueSpec)^(GenDecl)^(DeclStmt)^(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),y
+z,(Ident)^(ValueSpec)^(GenDecl)^(DeclStmt)^(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(BasicLit),2
+z,(Ident)^(ValueSpec)^(GenDecl)^(DeclStmt)^(BlockStmt)_(ReturnStmt)_(Ident),z
+int,(Ident)^(ValueSpec)^(GenDecl)^(DeclStmt)^(BlockStmt)_(AssignStmt:=)_(Ident),z
+int,(Ident)^(ValueSpec)^(GenDecl)^(DeclStmt)^(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),x
+int,(Ident)^(ValueSpec)^(GenDecl)^(DeclStmt)^(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),y
+int,(Ident)^(ValueSpec)^(GenDecl)^(DeclStmt)^(BlockStmt)_(AssignStmt:=)_(BinaryExpr:+)_(BasicLit),2
+int,(Ident)^(ValueSpec)^(GenDecl)^(DeclStmt)^(BlockStmt)_(ReturnStmt)_(Ident),z
+z,(Ident)^(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),x
+z,(Ident)^(AssignStmt:=)_(BinaryExpr:+)_(ParenExpr)_(BinaryExpr:+)_(Ident),y
+z,(Ident)^(AssignStmt:=)_(BinaryExpr:+)_(BasicLit),2
+z,(Ident)^(AssignStmt:=)^(BlockStmt)_(ReturnStmt)_(Ident),z
+x,(Ident)^(BinaryExpr:+)_(Ident),y
+x,(Ident)^(BinaryExpr:+)^(ParenExpr)^(BinaryExpr:+)_(BasicLit),2
+x,(Ident)^(BinaryExpr:+)^(ParenExpr)^(BinaryExpr:+)^(AssignStmt:=)^(BlockStmt)_(ReturnStmt)_(Ident),z
+y,(Ident)^(BinaryExpr:+)^(ParenExpr)^(BinaryExpr:+)_(BasicLit),2
+y,(Ident)^(BinaryExpr:+)^(ParenExpr)^(BinaryExpr:+)^(AssignStmt:=)^(BlockStmt)_(ReturnStmt)_(Ident),z
+2,(BasicLit)^(BinaryExpr:+)^(AssignStmt:=)^(BlockStmt)_(ReturnStmt)_(Ident),z
